# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Omnidapter Server",
    "description": "Self-hosted REST API wrapping the Omnidapter calendar integration library",
    "version": "0.3.1"
  },
  "paths": {
    "/v1/providers": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "List Providers",
        "operationId": "list_providers",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_ProviderMetadata__"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyAuth": []
          }
        ]
      }
    },
    "/v1/providers/{provider_key}": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "Get Provider",
        "operationId": "get_provider",
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "provider_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ProviderMetadata_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connections": {
      "post": {
        "tags": [
          "connections"
        ],
        "summary": "Create Connection",
        "description": "Create a new connection and begin the OAuth flow.",
        "operationId": "create_connection",
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnectionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CreateConnectionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "connections"
        ],
        "summary": "List Connections",
        "operationId": "list_connections",
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "name": "external_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "External Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_ConnectionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connections/{connection_id}": {
      "get": {
        "tags": [
          "connections"
        ],
        "summary": "Get Connection Endpoint",
        "operationId": "get_connection",
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ConnectionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "connections"
        ],
        "summary": "Delete Connection",
        "operationId": "delete_connection",
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/connections/{connection_id}/reauthorize": {
      "post": {
        "tags": [
          "connections"
        ],
        "summary": "Reauthorize Connection",
        "operationId": "reauthorize_connection",
        "security": [
          {
            "APIKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReauthorizeConnectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ReauthorizeConnectionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_ConnectionResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ConnectionResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/RequestMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[ConnectionResponse]"
      },
      "ApiResponse_CreateConnectionResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreateConnectionResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/RequestMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[CreateConnectionResponse]"
      },
...
      "ApiResponse_ProviderMetadata_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProviderMetadata"
          },
          "meta": {
            "$ref": "#/components/schemas/RequestMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[ProviderMetadata]"
      },
      "ApiResponse_ReauthorizeConnectionResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReauthorizeConnectionResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/RequestMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[ReauthorizeConnectionResponse]"
      },
...
      "ApiResponse_list_ProviderMetadata__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProviderMetadata"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/RequestMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[list[ProviderMetadata]]"
      },
...
      "AuthKind": {
        "type": "string",
        "enum": [
          "oauth2",
          "api_key",
          "basic"
        ],
        "title": "AuthKind",
        "description": "Supported authentication kinds."
      },
...
      "ConnectionConfigField": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "text"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": true
          },
          "placeholder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placeholder"
          },
          "example": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Example"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ConnectionConfigField",
        "description": "Describes a required/optional field for provider-specific connection config."
      },
      "ConnectionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "granted_scopes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Granted Scopes"
          },
          "provider_account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Account Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "provider",
          "external_id",
          "status",
          "status_reason",
          "granted_scopes",
          "provider_account_id",
          "created_at",
          "last_used_at"
        ],
        "title": "ConnectionResponse"
      },
...
      "CreateConnectionRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "redirect_url": {
            "type": "string",
            "title": "Redirect Url"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "redirect_url"
        ],
        "title": "CreateConnectionRequest"
      },
      "CreateConnectionResponse": {
        "properties": {
          "connection_id": {
            "type": "string",
            "title": "Connection Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "authorization_url": {
            "type": "string",
            "title": "Authorization Url"
          }
        },
        "type": "object",
        "required": [
          "connection_id",
          "status",
          "authorization_url"
        ],
        "title": "CreateConnectionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "pagination"
        ],
        "title": "ListMeta"
      },
      "ListResponse_ConnectionResponse_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ConnectionResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/ListMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ListResponse[ConnectionResponse]"
      },
      "OAuthMetadata": {
        "properties": {
          "authorization_endpoint": {
            "type": "string",
            "title": "Authorization Endpoint"
          },
          "token_endpoint": {
            "type": "string",
            "title": "Token Endpoint"
          },
          "supports_pkce": {
            "type": "boolean",
            "title": "Supports Pkce",
            "default": false
          },
          "scope_groups": {
            "items": {
              "$ref": "#/components/schemas/OAuthScopeGroup"
            },
            "type": "array",
            "title": "Scope Groups",
            "default": []
          },
          "default_scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Default Scopes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "authorization_endpoint",
          "token_endpoint"
        ],
        "title": "OAuthMetadata",
        "description": "OAuth-specific metadata for a provider."
      },
      "OAuthScopeGroup": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "scopes"
        ],
        "title": "OAuthScopeGroup",
        "description": "A named group of OAuth scopes."
      },
...
      "PaginationMeta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "PaginationMeta"
      },
      "ProviderMetadata": {
        "properties": {
          "provider_key": {
            "type": "string",
            "title": "Provider Key"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/ServiceKind"
            },
            "type": "array",
            "title": "Services"
          },
          "auth_kinds": {
            "items": {
              "$ref": "#/components/schemas/AuthKind"
            },
            "type": "array",
            "title": "Auth Kinds"
          },
          "oauth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OAuthMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "capabilities": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Capabilities",
            "default": {}
          },
          "connection_config_fields": {
            "items": {
              "$ref": "#/components/schemas/ConnectionConfigField"
            },
            "type": "array",
            "title": "Connection Config Fields",
            "default": []
          },
          "extra": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "provider_key",
          "display_name",
          "services",
          "auth_kinds"
        ],
        "title": "ProviderMetadata",
        "description": "Full provider metadata for introspection."
      },
      "ReauthorizeConnectionRequest": {
        "properties": {
          "redirect_url": {
            "type": "string",
            "title": "Redirect Url"
          }
        },
        "type": "object",
        "required": [
          "redirect_url"
        ],
        "title": "ReauthorizeConnectionRequest"
      },
      "ReauthorizeConnectionResponse": {
        "properties": {
          "connection_id": {
            "type": "string",
            "title": "Connection Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "authorization_url": {
            "type": "string",
            "title": "Authorization Url"
          }
        },
        "type": "object",
        "required": [
          "connection_id",
          "status",
          "authorization_url"
        ],
        "title": "ReauthorizeConnectionResponse"
      },
...
      "RequestMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "RequestMeta"
      },
      "ServiceKind": {
        "type": "string",
        "enum": [
          "calendar",
          "crm"
        ],
        "title": "ServiceKind",
        "description": "Top-level service categories."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyAuth": {
        "type": "http",
        "description": "Main Server API Key (omni_*)",
        "scheme": "bearer"
      },
      "LinkTokenAuth": {
        "type": "http",
        "description": "Connect UI session token (cs_*). The one-time lt_* bootstrap token is only accepted by POST /connect/init.",
        "scheme": "bearer"
      }
    }
  }
}